Convert typography component to ts remove theme
Why:
Currently the Typography component is implemented with significant logic
surrounding being able to allow the user to update the theme.
Additionally the component defines a few typography style classes
inline. Given that we do not want to support dynamic themes and would
like to have all of the styles consistently handled by the design
system, we would like to refactor the Typography component such that it
is only responsible for handling right to left language adjustments.

We would also like to have this component be covered by the ts compiler.

This commit:
Updates the Typography component to use type script, removes unused
logic, and uses styles from the common project styles. While converting
to ts, a handful of call sites were discovered to have typos and type
errors, which were also fixed.