Tp remove bte naming build fix (#1027)
* Replace BTE with BT within application

WHY:
Initially, BTE was used even though it didn't stand for anything. I think this came from
BLE, but got lost in translation. Now the PathCheck apps are GPS and BT, so removing
use of BTE within the app will hopefully unify the language and remove the use of BTE
in conversations.

* Update package names to fix build.

Co-authored-by: Thomas Paresi <thomas.paresi@thoughtbot.com>
Co-authored-by: Dick Lucas <dick@fullsend.io>