Build exposure history from raw exposures
Why:
We would like to be creating an exposure history for the user based off
of data received from the Exposure Notification Server data which is
passed to the JavaScript layer from the native layer.

This commit:
Introduces the logic for converting a list of raw Exposures sent from
the native side to an ExposureHistory. Additionally we refactored the
NativeModule a bit and exposed the debug screen to the production build
so that users can test in test flight.

Currently we are not receiving data from the GAEN server, but we do have
mocked data coming from the native layer. A user can tap 'simulate
exposure' on the debug screen to create a new simulated exposure.