Add safe area view for navigation
Why:
Currently the navigation ui is not set up to respect safe area for large
format iPhones. Additionally the status bar barStyle does not update for
dark and light backgrounds declaratively. We would like for the
navigation screens to look consistent and professional on all phone
formats.

This commit:
Introduces a safe area view that is compatible with react navigation.
Given the JS > native approach that the react-navigation team takesj, the
safe area view that is provided directly by react-native does not
integrate with the navigator update cycle as one happens on the native
side and the other on the js side.

A similar js implementation is provided for setting the status bar, its
unfortunately tied to a side effect, but it seems to be the only
available solution compatible with react-navigation at this point.

The various navigation screens were updated to use these new components
and side effect via a `useStatusBarEffect` custom hook. The navigation
bar wrapper was also updated to use the correct styles per the current
design.