Show future days on calendar
Why:
Currently we would like to show days in the future on the calendar such
that the day labels line up with the correct day number.

This commit:
Introduces the logic to build the correct days in the ExposureHistory
list, that is, we determine when the next saturday is and pad out the
history such that the last day in the history is the up coming
staturday.

We separated out the concept of building an ExposureHistory as a list,
and building out the ExpousreHistory as a Map of days to exposures,
which we are calling an ExposureInfo. This will allow us flexibility in
presenting the exposure history in any calendar format a bit more easily
moving forward, for example other communities may not use the gregorian
calendar or we may discover that users would benefit from not seeing
this info in a calendar view at all.

This has a consequence to any tracing strategy that wants to be consumed
by the calendar view ui, as it is expected to now provide an
ExposureInfo rather than an ExposureHistory.