Add strategy content to tracing strategy
Why:
We would like to include all of the strategy specific content assets in
the tracingStrategy object provided to the app on app load. This will
allow us to not need to check the strategy for each individual piece of
copy or asset as we simply provide all of the strategy dependent content
in the root component.

This commit:
Removes the useAssets hook and replaces its usage with a
useStrategyContent hook which gets is data from a prop passed to the
Tracing Strategy provider on mounting of the root component.
Additionally we add the test factories library, fishery, to provide test
factories for the specs related to this behavior.

Note that a better strategy to handle the translated content may be to
have separation at the translation json files, rather than at the
javascript object level. A future commit may want to introduce this to
replace the `StrategyCopy` portion of this workflow.