Introduce tracing strategy provider (#1081)
Why:
Currently we do not have much protection against one version of the app
accessing functionality intended to only be used another version. For
example, in the BT build of the app, any component could request
location permissions, which would cause the app to error. Given that
situations like these could be grounds for violating compliance, we
would like to introduce a better strategy for providing tracing strategy
specific logic and functionality.

Additionally, the application has many conditional checks of the tracing
strategy scattered throughout the code base (`isGPS` and `!isGPS`). This
is prone to human error, is likely to be forgotten, and introduces
significant cognitive noise while reading the implementation of a
feature. Given that near term product goals include introducing more
tracing strategies, this is also prime for a combinatoric explosion of
complexity. We would like to reduce the number of sites in which we need
to be conditionally checking the tracing strategy, ideally to only one.

This commit:
Introduces the concept of a TracingStrategy as an interface, which can
be used as the common schema to which all tracing strategies can be
built against. And which UI application code can rely on to provide a
functional user interface for that tracing strategy.

We check the tracing strategy in the root App component and provide the
strategy specific functionality to the rest of the application via a
context. This removes the need of child components to need to check the
current strategy and makes it inconvenient and unlikely for a child
component to access functionality belonging to a different strategy.

Note that this commit does not remove every instance of isGPS from the
code base to reduce the size of this commit. Future commits should
include introducing the strategy specific content, assets, and routes to
the tracing strategy interface.

Finally by having a notion of a TracingStrategy in the type system it
becomes significantly easier to reason about the portions of the which
are different between the strategies and ideally we will be able to keep
the interface as small as possible to maximize the amount of common code
between the strategies.