Display en device status on home screen (#1067)
Why:
currently the app will correctly enable exposure notifications on the
users device, but it will not correctly show the device status in the
app. We would like for the Home screen to reflect whether or not the
device has exposure notifications enabled, and if they are not we would
like to render a button that will request to enable permissions on the
device.

This commit:
Introduces a Home Screen component to the tracingStrategy which for the
bt build shows a new HomeScreen component and for the gps build will
render the existing MainScreen. Ideally in a future commit, we can
consolidate these home screens, but for sake of time it was easier to
introduce a new component.

Co-Authored-By: mattthousand <matt@nicethings.io>
Co-Authored-By: johnschoeman <johnschoeman1617@gmail.com>
Co-Authored-By: tparesi <thomas.paresi@thoughtbot.com>