Fix start affected user flow crash
Why:
Currently the app will crash if a user tries to start the affected user
flow while there are not health authorities in the store.

This commit:
Provides a fake health authority if there are none such that the app
will not crash. A future commit should properly handle this null state
and we should remove this fake authority.