Remove theme from onboarding screens
Why:
Currently the onboarding screens include a no longer wanted theme
provider. Given that we actively do not want to allow for dynamic theme
changes in the app we would like to remove the theme component.

This commit:
Refactors the onboarding components such that we can remove the theme
component while keeping the same visual design. We ended up needing to
remove the permissions screens from the purview of the
`OnboardingTemplate` as the permissions screens are significantly
different from the onboarding content screens in terms of markup, view
logic, and style so as to keep the onboarding template for them would
require introducing significant unnecessary complexity and ultimately
make the screens less amenable to future changes in requirements.

Note that this is entirely a refactor and no new functionality has been
added.