Convert button.js to tsx and remove theme (#1132)
Why:
We would like remove unused code and have as much of the app covered by
static analysis as possible. Currently the Button component is a heavily
used component that has a dependency on the ThemeProvider and does not
have type coverage. We would like to remove the dependency on the theme
provider and convert it to typescript.

This commit:
Converts Button.js to Button.tsx, removes the theme and converts a few
call sites of <Button /> to a simpler implementation of
<TouchableOpacity />.

Note that we likely want to remove Button.tsx altogether in a future
commit as it doesnt provide much value and it has a maintenance cost. We
are not removing it altogether in this commit to reduce the size of the
commit and for sake of time.