Remove gitsubmodules for env
Why:
A previous commit introduces a gitsubmodules strategy for updating and
.env files, which broke builds for prs from forked repos. Additionally
its possible to handle environment variables in a more 12-factored app
approach and only use an explicit .env file for local development.

This commit:
Reverts the previous commit, removes the .env.release and .env.staging
files from version control and updates the github action workflows to
read the env variables from the machine env rather than a file in the
repo.