Move eula checkbox to welcome screen (#1135)
Why:
Currently the user flow for the initial welcome screen requires the user
to open the terms modal and then tap an agreement checkbox. We would
like to move the checkbox to the welcome screen itself and have the
modal only contain the terms content, per the current designs.

This commit:
Refactors Onboarding1 and EulaModal screen such that the checkbox can be
moved. While doing so we changed the name of Onboarding1 to Welcome and
converted the screen to typescript. We also updating some of the styling
layout and brought the pill button style into the main styles folder.

Co-Authored-By: ericwbailey <eric.w.bailey@gmail.com>