Request notification permissions for bt app
Why:
We would like for the bt build of the app to request notification
permissions during onboarding as they are necessary for the Exposure
Notification flow.

This commit:
Adds the screen to the onboarding flow for the BT build. Note that we
needed to duplicate the component from the GPS build. Ideally a future
commit will refactor the flow such that it can be agnostic to the
tracing strategy and we can avoid this code duplication.

Co-Authored-By: mattthousand <matt@nicethings.io>