Introduce affected user flow for bt (#1174)
Why:
Currently there is a handful of bugs in the Affected User flow (called
Export in the app currently) for the BT version. This is largely due to
the fact that the Export Flow was built with the assumption that the
user would be selecting the HealthAuthority as is the case with the GPS
version. We would like to update the Screens to be agnostic to this,
however for sake of time, we are introducing a similar flow for the BT
version and mounting the correct flow based off of the tracing Strategy.

This commit:
Introduces the AffectedUserFlow for the BT app. And moves the Export
flow into the gps folder, as it is no longer a shared component.

Note that this commit moved and copied all of the files for the Export
flow so that the git diff looks quite large compared to the actual
changes that were introduced to the bluetooth side. The content and user
flow largely stayed the same, we updated where the HealthAuthority info
and network calls happened for the bt side and changed the code length
from 6 to 8. We changed no functionality on the GPS side, we simply
moved the files to the gps folder.

Co-Authored-By: alejandrodustet <aledustet@gmail.com>