Fix cyclic dependency warning
Why:
----
Since the store needs to be made available when checking for intersects
for the health authority access, there is a cyclic dependency warning
that shows up every time the application builds. We need a way to avoid
this cyclic dependency from happening.

How:
----
The approach was to add a getter and a setter singleton to access the
store via a third file from `Intersect.js`. A possible future approach
could be to move the intersect logic onto the redux store but it is
future work to be done since it will involve a more in-depth rework of
the logic.

This Commit:
----
- Add a `StoreAccessor` to make the store available without importing
the actual store.
- Use the StoreAccessor from the `Intersect` helper

[trello](https://trello.com/c/Cb9wAu2f)

Co-Authored-By: Jacob Jaffe <jacobmaxjaffe@gmail.com>