Add react testing library preset to allow using matchers
Why:
----
When we are running our tests, most of the components that get rendered
into a screen are `native` powered, making it difficult for test
renderers to display them in a proper way. There are a set of presets
that make this possible. In an effort to make our integration and unit
tests more similar to what the user is seeing and giving us more
confidence in what we can test, we need to add the react native testing
library preset that will allow us to run matchers on what is being
displayed to the user on the screen. This way our tests can be more
similar to the actual behavior of the application.

This Commit:
----
- Add preset `@testing-library/react-native` to jest config
- Update snapshots for the new renderer result
- Stop babel from transforming modules when the test runner is working