Fetch verification certificate
Why:
After verifying the users provided code to the verification server and
receiving a token, we need to generated a hashed version of the users
keys (and potentially metadata) and post this to verification server to
receive a certificate which we can use to post exposure keys to the key
server.

This commit:
Introduces the logic for generating and HMAC of the current exposure
keys and posting this data along with the previously received token to
api/certificate.

We introduced react-native-simple-crypto for calculating the HMAC
hashing on the JS side as this will reduce the amount of code that the
native layer will need to implement. In a future iteration, it might
make sense to do all of the cryptographic logic in the native layer as
this will allow us to remove the dependency react-native-simple-crypto

A native module promise, `ExposureKeyModule.fetchExposureKeys`, was
added to fetch the exposure keys and serialize them to send the
encrypted payload to the verification server.

Next Steps:
- Save the hmacKey to `realm`
- Save the certificate generated on the verification server to `realm`
- Post the exposure keys to the `GAEN` with the certificate and the HMAC
key
- Handle errors on the exposure keys, post exposure data flow
- Add in user metadata to the request that goes to the Verification
Server

Co-Authored-By: Alejandro Dustet<aledustet@gmail.com>