Publish keys with cert to exposure key server
Why:
We need to send the verification credentials generated by the
verification server to the exposure key server so that the exposure keys
provided by the device can be cross verified by the servers. This
process requires communication from the verification server flow to the
exposure key server flow.

This commit:
Adds the generated hmac key and certificate, added in a prior commit
PR:1223 , to the submitDiagonsisKeys method on the native module. We
updated the PublishConsentScreen to handle delegating this to the Native
Layer. We also removed the stubbed logic for saving the HMAC key to
realm as this is not required for the at this point.

Co-Authored-By: Alejandro Dustet <alejandro@thoughtbot.com>