Avoid running Intersect logic on BT flavor
Reviewers:
----
This problem was brought to our attention because of a warning that was
showing up every time we were building the app with the `bt`
configuration. This was the error we were seeing:

```
LOG  [intersect] tick entering on iOS;
WARN  Possible Unhandled Promise Rejection (id: 0):
TypeError: null is not an object (
  evaluating '_reactNative.NativeModules.SecureStorageManager.getLocations'
)
```

This happens because the native module responsible for that logic
`SecureStorageManager` is not available on the `bt` build.
We corrected the first occurrence of this error, happening on app load
after checking that the app was on the `gps` configuration before
starting the background service. After we closed the app, the error kept
showing up, because the middleware `onChangedSelectedHealthAuthorities`
was also starting a check on the intersect service.

Why:
----
The store applies a middleware to check for changes on the health
authority and then it runs the `IntersectService` pertinent to check for
exposures with the gps detection logic. This middleware gets applied
regardless of what the logic is in the app and we are seeing errors
because the native module used for this purpose is not available on the
`bt` build.

This PR:
----
- Move the creation of the store to a method to accept middleware as an
option
- Add middleware to the tracing strategy