Script to update display name based on environment
Why:
----
Each HA will have different settings for our builds, one of those is the
display name. We need a way to update those values for each one of the
builds.

This Commit:
----
- Extend setup env script to update the display name of the android and
the ios app for the bt builds

Reviewers:
----
This script is counting on having `xmllint` and `PlistBuddy` in the
executing environment. Given that builds need to happen on `MacOS` those
two libraries are part of the system. The solution is very rigid but it
is a start on how we might tackle the configurations per HA.