Fetch ha specific config with curl (#1283)
Why:
Currently we are fetching ha specific environments with git submodules,
this causes ci builds to break for contributions that made from outside
contributors via forks. We would like to be able to quickly update the
HA config and also allow contributions from individuals outside of the
github organization.

This commit:
Removes the gitsubmodules and replaces them with a `bin/fetch_ha_env`
script that uses curl over the github api. Developers that want to use
this script will need to set a GITHUB_PERSONAL_TOKEN in their local
.env, which was also added in this pr.

So we are introducing a distinction between `.env` and `.env.bt` where
`.env` is local development environment variables and `.env.bt` are
variables meant to be used in the built binaries for the app.